Implement BeatmapSearchMultipleSelectionFilterRow
