+++ osu.Game/Overlays/BeatmapListing/BeatmapSearchFilterRow.cs
@@ -1,20 +1,16 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
 using JetBrains.Annotations;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
-using osu.Framework.Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
 using osuTK;
-using osuTK.Graphics;
 using Humanizer;
 using osu.Game.Utils;
 
@@ -98,7 +94,7 @@ namespace osu.Game.Overlays.BeatmapListing
 
             protected override Dropdown<T> CreateDropdown() => new FilterDropdown();
 
-            protected override TabItem<T> CreateTabItem(T value) => new FilterTabItem(value);
+            protected override TabItem<T> CreateTabItem(T value) => new FilterTabItem<T>(value);
 
             private class FilterDropdown : OsuTabDropdown<T>
             {
@@ -116,64 +112,6 @@ namespace osu.Game.Overlays.BeatmapListing
                     }
                 }
             }
-        }
-
-        protected class FilterTabItem : TabItem<T>
-        {
-            protected virtual float TextSize => 13;
-
-            [Resolved]
-            private OverlayColourProvider colourProvider { get; set; }
-
-            private readonly OsuSpriteText text;
-
-            public FilterTabItem(T value)
-                : base(value)
-            {
-                AutoSizeAxes = Axes.Both;
-                Anchor = Anchor.BottomLeft;
-                Origin = Anchor.BottomLeft;
-                AddRangeInternal(new Drawable[]
-                {
-                    text = new OsuSpriteText
-                    {
-                        Font = OsuFont.GetFont(size: TextSize, weight: FontWeight.Regular),
-                        Text = CreateText(value)
-                    },
-                    new HoverClickSounds()
-                });
-
-                Enabled.Value = true;
-            }
-
-            protected virtual string CreateText(T value) => (value as Enum)?.GetDescription() ?? value.ToString();
-
-            [BackgroundDependencyLoader]
-            private void load()
-            {
-                updateState();
-            }
-
-            protected override bool OnHover(HoverEvent e)
-            {
-                base.OnHover(e);
-                updateState();
-                return true;
-            }
-
-            protected override void OnHoverLost(HoverLostEvent e)
-            {
-                base.OnHoverLost(e);
-                updateState();
-            }
-
-            protected override void OnActivated() => updateState();
-
-            protected override void OnDeactivated() => updateState();
-
-            private void updateState() => text.FadeColour(Active.Value ? Color4.White : getStateColour(), 200, Easing.OutQuint);
-
-            private Color4 getStateColour() => IsHovered ? colourProvider.Light1 : colourProvider.Light3;
         }
     }
 }
